Remove duplicate or obselete idv specs
**Why**: The specs in this commit have either been moved into idv step
specs, or are moved by this commit. This commit also reshuffles a few
existing specs to make sure that our coverage at the feature level does
not drop.
